refactor: organize typing practice
