+++ WT4Q/src/app/tools/typing-practice/utilities/wordBank.ts
@@ -1,0 +1,19 @@
+export const WORDS_EASY =
+  'the of and to a in is you that it he was for on are as with his they I at be this have from or one had by word but not what all were we when your can said there use an each which she do how their if will up other about out many then them these so some her would make like him into time has look two more write go see number no way could people my than first water been call who oil its now find long down day did get come made may part'.split(/\s+/);
+
+export const WORDS_MEDIUM =
+  'bridge motion elegant figure silver window forest travel storm gentle whisper bright velvet hidden puzzle lantern shadow crystal rhythm galaxy vessel anchor meadow distant silent fragile frequent fabric hollow vivid subtle ripple orbit spiral crater amber vivid moment narrow velvet luminous shimmer cascade swift granite magnet fossil falcon maple cedar salmon azure sprint dodge arrow quiver'.split(/\s+/);
+
+export const WORDS_HARD =
+  'quixotic zephyr labyrinthine juxtapose chrysanthemum vestibule silhouette nebulous quintessential panacea vicissitude synecdoche susurrus chiaroscuro phantasmagoria incandescence mellifluous grandiloquent indefatigable perspicacious sesquipedalian ebullient paucity verisimilitude apotheosis intransigent obfuscate recondite anachronistic obstreperous'.split(/\s+/);
+
+export const QUOTES = [
+  'The quick brown fox jumps over the lazy dog.',
+  'Programs must be written for people to read, and only incidentally for machines to execute.',
+  'Premature optimization is the root of all evil.',
+  'Simplicity is the soul of efficiency.',
+  'Talk is cheap. Show me the code.',
+  'Any fool can write code that a computer can understand. Good programmers write code that humans can understand.',
+  'First, solve the problem. Then, write the code.',
+  'In theory, theory and practice are the same. In practice, they\u2019re not.',
+];
